[WIP][Extension][Spark] New Authz Plan Serde Layer
